DBSDataContainer and data fetch fix (replace ints with longs)
Former-commit-id: 9927b909711056d3604f185bd9f74181512f8995

+++ org.jkiss.dbeaver/core/src/org/jkiss/dbeaver/model/dbc/DBCStatement.java
@@ -56,7 +56,7 @@ public interface DBCStatement extends DBPObject, DBRBlockingObject
 
     void close();
 
-    void setLimit(int offset, int limit) throws DBCException;
+    void setLimit(long offset, long limit) throws DBCException;
 
     /**
      * Statement data container.
